requirements.txt
openai==1.101.0
python-dotenv==1.1.1
requests==2.32.5
Flask==3.0.3
flask-cors==4.0.0
